feat: move User model to User entity, connect User to DB
