+++ src/common/DbTables/UserTable.ts
@@ -1,4 +1,4 @@
-import User from '../../resources/users/user.model';
+import User from '../../resources/users/user.entity';
 import Table from './Table';
 
 class UserTable extends Table<User> {}
