+++ src/common/inMemoryDb.ts
@@ -1,4 +1,4 @@
-import User from '../resources/users/user.model';
+import User from '../resources/users/user.entity';
 import Board from '../resources/boards/board.model';
 import Task from '../resources/tasks/task.model';
 import UserTable from "./DbTables/UserTable";
